pyproject.toml
[tool.poetry]
name = "pyguiadapter-i18n"
version = "0.1.0"
description = ""
authors = ["zimolab <zimolab@aliyun.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.14"


pyguiadapter = "^1.0.5"
python-i18n = "^0.3.9"
pyqt6 = "^6.7.1"
[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"